Fix question text output
The original files are encoded in ISO-8859 and are retrieved as literal
bytestreams, but as a result, trying to construct strings out of them
results in the repr version of the bytestrings. Calling the decode()
function instead fixes that.

In addition, the whitespace-split version of the question text is being
written to the output file, but as a result, all newlines are
obliterated. Write the not-whitespace-split version of the question text
to the file instead.