More fixing up of the question splitting logic
Previous version used '[0-9]+\.' as the regex pattern for splitting up
questions, but that could potentinally result in individual questions be
split in half if a sentence in the question ended with a numeral.
Instead, use regex pattern '\n *[0-9]+\.' to ensure that the number is
at the beginning of the line.

In addition, to account for the fact that the original question number
is lost due to the splitting process, the appropriate number of spaces
are added to each question printed to keep the question lines aligned.

Also, in some exams, Mackey has two multiple choice sections. Code has
been adjusted to account for that, including proper numbering and not
accidentally including Mackey's Multiple Choice grading rubric as part
of the 1st question in Multiple Choice part 2.